Fix compile for 64-bit size_t compatibility

+++ Matrix/ConvolverFactory.cc
@@ -123,7 +123,7 @@ vector<double> TemplateConvolverFactory::calcKernel(unsigned int i) const {
     vector<double> vs(i);
     assert(s0 >= 0 && s0 < (int)v.size());
     for(int s=s0; s>=std::max(0,int(s0-i+1)); s--) vs[s0-s] = v[s];
-    for(int s=s0+1; s<(int)std::min(v.size(),i); s++) vs[i-(s-s0)] = v[s];
+    for(int s=s0+1; s<(int)std::min(v.size(),(size_t)i); s++) vs[i-(s-s0)] = v[s];
     return vs;
 }
